Parse entire modulefile for rpaths
o Previously, the assumption was made that the first library path found in a
compiler modulefile was the location of all dependent libraries.  This is a bad
assumption, as many HPC clusters have installed compilers with dependent shared
libraries distributed over many directories (not necessarily co-located).  This
change parses the entire file, and returns a colon-delimited string with all the
found paths.